requirements.txt
nonebot2>=2.0.0
nonebot-adapter-onebot>=1.0.0
openai>=1.0.0
httpx>=0.25.0
pyyaml>=6.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
chromadb>=0.4.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
